requirements.txt
Werkzeug==0.16.1
tensorflow==2.4.0
opencv-python==4.5.1.48
easydict==1.9
Pillow==8.2.0
requests>=2.9.1
pixellib==0.4.8
onnxruntime==1.8.1
onnx==1.10.1
scikit-image==0.16.2
tensorflow-model-optimization==0.7.0
tf2onnx==1.9.2
pytorch=1.10.0
thop